Move plan/ to query_planner/
