+++ nidx/src/searcher.rs
@@ -16,10 +16,6 @@
 pub mod grpc;
 mod index_cache;
 mod query_language;
-mod plan {
-    pub mod prefilter;
-    pub mod suggest;
-}
 mod query_planner;
 mod shard_merge;
 mod shard_search;
